chore(workspace): enforce UX isolation — the condition builder becomes a kit module
The nested filter builder over the published condition AST was app-local for no
reason but where it happened to be written. Its logic imported an API contract
type for the icon it drives and a message-key union for the labels it drew;
neither is anything the builder needs to know.

It now lives in @stll/workspace-ui/conditions and resolves no strings at all. A
host describes its operands as FieldOptions and brings every label the builder
draws, including its operators — the message-key table stays in the app, where
the guard proving each key exists still bites. The formula editor is a slot,
since its expression language belongs to the host, and the date editor takes
the locale and its three labels as props.

Nothing at the six call sites moves: the app's builder is a wrapper that
supplies this app's catalogue and formatting locale.

Boy-scout: the workspace kit is now swept by the two key-hygiene rules its
components already carried disables for.
